BookStoreApi-master/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.11.0
dnspython==2.0.0
drf-yasg==1.17.1
future==0.18.2
greenlet==0.4.16
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
monotonic==1.5
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
pycparser==2.20
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zope.interface==5.1.0
